Refactor GraphQL types to have same file format as queries and mutations.
