+++ graphql/queries/session.js
@@ -4,7 +4,7 @@ import {
   GraphQLList
 } from 'graphql'
 
-import sessionType from '../types/SessionType'
+import sessionType from '../types/session'
 import SessionBase from '../../model/session/session.base'
 
 var sessionBase = new SessionBase()
