+++ graphql/types/user.js
@@ -5,7 +5,7 @@ import {
   GraphQLInt
 } from 'graphql'
 
-import sessionType from '../types/SessionType'
+import sessionType from '../types/session'
 import SessionBase from '../../model/session/session.base'
 
 let sessionBase = new SessionBase()
